+++ mwclient/__init__.py
@@ -23,12 +23,15 @@
  OTHER DEALINGS IN THE SOFTWARE.
 """
 
-from errors import *
-from client import Site, __ver__
-import ex
+from mwclient.errors import *
+from mwclient.client import Site, __ver__
+import logging
+import warnings
+
+# Show DeprecationWarning
+warnings.simplefilter('always', DeprecationWarning)
 
 # Logging: Add a null handler to avoid "No handler found" warnings.
-import logging
 try:
     from logging import NullHandler
 except ImportError: