+++ mwclient/errors.py
@@ -56,3 +56,16 @@ class NoSpecifiedEmail(EmailError):
 
 class NoWriteApi(MwClientError):
     pass
+
+
+class InvalidResponse(MwClientError):
+
+    def __init__(self, response_text=None):
+        self.message = 'Did not get a valid JSON response from the server. Check that ' + \
+                       'you used the correct hostname. If you did, the server might ' + \
+                       'be wrongly configured or experiencing temporary problems.'
+        self.response_text = response_text
+        MwClientError.__init__(self, self.message, response_text)
+
+    def __str__(self):
+        return self.message