+++ mwclient/listing.py
@@ -1,6 +1,9 @@
-import client
-import page
+import six
+import six.moves
 from six import text_type
+from mwclient.util import parse_timestamp
+import mwclient.page
+import mwclient.image
 
 
 class List(object):
@@ -22,7 +25,7 @@ class List(object):
         self.count = 0
         self.max_items = max_items
 
-        self._iter = iter(xrange(0))
+        self._iter = iter(six.moves.range(0))
 
         self.last = False
         self.result_member = list_name
@@ -31,15 +34,15 @@ class List(object):
     def __iter__(self):
         return self
 
-    def next(self, full=False):
+    def __next__(self, full=False):
         if self.max_items is not None:
             if self.count >= self.max_items:
                 raise StopIteration
         try:
-            item = self._iter.next()
+            item = six.next(self._iter)
             self.count += 1
             if 'timestamp' in item:
-                item['timestamp'] = client.parse_timestamp(item['timestamp'])
+                item['timestamp'] = parse_timestamp(item['timestamp'])
             if full:
                 return item
 
@@ -56,8 +59,12 @@ class List(object):
             self.load_chunk()
             return List.next(self, full=full)
 
+    def next(self, full=False):
+        """ For Python 2.x support """
+        return self.__next__(full)
+
     def load_chunk(self):
-        data = self.site.api('query', (self.generator, self.list_name), *[(text_type(k), v) for k, v in self.args.iteritems()])
+        data = self.site.api('query', (self.generator, self.list_name), *[(text_type(k), v) for k, v in six.iteritems(self.args)])
         if not data:
             # Non existent page
             raise StopIteration
@@ -76,11 +83,11 @@ class List(object):
 
     def set_iter(self, data):
         if self.result_member not in data['query']:
-            self._iter = iter(xrange(0))
+            self._iter = iter(six.moves.range(0))
         elif type(data['query'][self.result_member]) is list:
             self._iter = iter(data['query'][self.result_member])
         else:
-            self._iter = data['query'][self.result_member].itervalues()
+            self._iter = six.itervalues(data['query'][self.result_member])
 
     def __repr__(self):
         return "<List object '%s' for %s>" % (self.list_name, self.site)
@@ -88,7 +95,7 @@ class List(object):
     @staticmethod
     def generate_kwargs(_prefix, *args, **kwargs):
         kwargs.update(args)
-        for key, value in kwargs.iteritems():
+        for key, value in six.iteritems(kwargs):
             if value is not None and value is not False:
                 yield _prefix + key, value
 
@@ -130,15 +137,15 @@ class GeneratorList(List):
 
         self.result_member = 'pages'
 
-        self.page_class = page.Page
+        self.page_class = mwclient.page.Page
 
     def next(self):
         info = List.next(self, full=True)
         if info['ns'] == 14:
             return Category(self.site, u'', info)
         if info['ns'] == 6:
-            return page.Image(self.site, u'', info)
-        return page.Page(self.site, u'', info)
+            return mwclient.image.Image(self.site, u'', info)
+        return mwclient.page.Page(self.site, u'', info)
 
     def load_chunk(self):
         # Put this here so that the constructor does not fail
@@ -147,10 +154,10 @@ class GeneratorList(List):
         return List.load_chunk(self)
 
 
-class Category(page.Page, GeneratorList):
+class Category(mwclient.page.Page, GeneratorList):
 
     def __init__(self, site, name, info=None, namespace=None):
-        page.Page.__init__(self, site, name, info)
+        mwclient.page.Page.__init__(self, site, name, info)
         kwargs = {}
         kwargs['gcmtitle'] = self.name
         if namespace:
@@ -189,9 +196,9 @@ class PageList(GeneratorList):
         if self.namespace == 14:
             return Category(self.site, self.site.namespaces[14] + ':' + name, info)
         elif self.namespace == 6:
-            return page.Image(self.site, self.site.namespaces[6] + ':' + name, info)
+            return mwclient.image.Image(self.site, self.site.namespaces[6] + ':' + name, info)
         elif self.namespace != 0:
-            return page.Page(self.site, self.site.namespaces[self.namespace] + ':' + name, info)
+            return mwclient.page.Page(self.site, self.site.namespaces[self.namespace] + ':' + name, info)
         else:
             # Guessing page class
             if type(name) is not int:
@@ -199,11 +206,11 @@ class PageList(GeneratorList):
                 if namespace == 14:
                     return Category(self.site, name, info)
                 elif namespace == 6:
-                    return page.Image(self.site, name, info)
-            return page.Page(self.site, name, info)
+                    return mwclient.image.Image(self.site, name, info)
+            return mwclient.page.Page(self.site, name, info)
 
     def guess_namespace(self, name):
-        normal_name = page.Page.normalize_title(name)
+        normal_name = mwclient.page.Page.normalize_title(name)
         for ns in self.site.namespaces:
             if ns == 0:
                 continue
@@ -223,7 +230,7 @@ class PageProperty(List):
         self.generator = 'prop'
 
     def set_iter(self, data):
-        for page in data['query']['pages'].itervalues():
+        for page in six.itervalues(data['query']['pages']):
             if page['title'] == self.page.name:
                 self._iter = iter(page.get(self.list_name, ()))
                 return